DS-4440 GDPR - Anonymize statistics feature - move tests to dspace-api
