+++ dspace-api/src/test/data/dspaceFolder/config/spring/api/solr-services.xml
@@ -44,6 +44,8 @@
           class="org.dspace.statistics.MockSolrLoggerServiceImpl"
           lazy-init="true"/>
 
-    <bean class="org.dspace.statistics.MockSolrStatisticsCore" autowire-candidate="true"/>
+    <bean id="org.dspace.statistics.SolrStatisticsCore"
+          class="org.dspace.statistics.MockSolrStatisticsCore"
+          autowire-candidate="true"/>
 
 </beans>
